+++ libraries/FunctionGenerator/test/unit_test_001.cpp
@@ -33,6 +33,7 @@ unittest_setup()
   fprintf(stderr, "FUNCTIONGENERATOR_LIB_VERSION: %s\n", (char*) FUNCTIONGENERATOR_LIB_VERSION);
 }
 
+
 unittest_teardown()
 {
 }
@@ -46,8 +47,10 @@ unittest(test_constructor)
   assertEqualFloat(1, gen.getPeriod(), 0.0001);
   assertEqualFloat(1, gen.getFrequency(), 0.0001);
   assertEqualFloat(1, gen.getAmplitude(), 0.0001);
+
   assertEqualFloat(0, gen.getPhase(), 0.0001);
   assertEqualFloat(0, gen.getYShift(), 0.0001);
+  assertEqualFloat(50, gen.getDutyCycle(), 0.0001);
 }
 
 
@@ -60,8 +63,10 @@ unittest(test_constructor_param)
   assertEqualFloat(4.00, gen.getPeriod(), 0.0001);
   assertEqualFloat(0.25, gen.getFrequency(), 0.0001);
   assertEqualFloat(2.50, gen.getAmplitude(), 0.0001);
+
   assertEqualFloat(0.33, gen.getPhase(), 0.0001);
   assertEqualFloat(-0.5, gen.getYShift(), 0.0001);
+  assertEqualFloat(50,   gen.getDutyCycle(), 0.0001);
 }
 
 
@@ -85,6 +90,9 @@ unittest(test_set_get_param)
 
   gen.setYShift(6.89);
   assertEqualFloat(6.89, gen.getYShift(), 0.0001);
+
+  gen.setDutyCycle(14.28);
+  assertEqualFloat(14.28, gen.getDutyCycle(), 0.0001);
 }
 
 
@@ -112,6 +120,9 @@ unittest(test_line_zero)
   TODO wave forms 6x ?
 */
 
+
 unittest_main()
 
-// --------
+
+//  -- END OF FILE --
+